+++ src/clev2er/algorithms/seaice_stage_2/tests/test_alg_grid_for_volume.py
@@ -128,7 +128,9 @@ def test_grid_for_volume(
     try:
         l1b = Dataset(l1b_sar_file)
         logger.info("Loaded %s", l1b_sar_file)
-        f_time = datetime.fromtimestamp(np.min(l1b["measurement_time"])).strftime("%Y%M")
+        f_time = datetime.fromtimestamp(np.min(l1b["measurement_time"]).astype(int)).strftime(
+            "%Y%M"
+        )
         grid_file_name = f"{f_time}_grids.nc"
     except IOError:
         assert False, f"{l1b_sar_file} could not be read"
